Added setPixelSnappedFrame and setPixelSnappedCenter methods.
If you called setXOrigin and setYOrigin in the same animation block the
call to setXOrigin would be clobbered by the call to setYOrigin because
the both modify the fame.

Two new methods have been added that will allow you to set the frame or
center of a UIView all at once so you can modify multiple values and
not break your animations.
These methods will also automatically floor the needed values to
prevent subpixel rendering.